refactor: rename materiality service files to naming convention
